Implement skip parameter for inf-test-reduce-replay
this allows, instead of going forward one-by-one, to go in steps of
more than one request. This is useful for very big records, where one first
needs to get close to where the problem is and then one can carry on in
smaller steps to get the most minimal record possible.