InfTextUndoGrouping: Don't get stuck with transforming too old requests
It could happen that a request was transformed over tens of thousands of
other requests if a user waits a long time between two keystrokes, and
there was much activity by others between the two keystrokes. In such a case,
simply do not group two requests which are very far apart from each other.