Remove error parameter from inf_communication_object_received()
Implement this by adding the InfSession::error signal in InfSession.
This allows client code to handle the error in whatever way is
approriate, instead of just logging it to stderr. InfSession itself does
not know about the nature of the error, or whether it is fatal or not.
For example, if a request cannot be processed it would be fatal for a
client (or a peer-to-peer participant) but not for a server (infinoted).
In the server case, it would simply ignore the request, not forward it
to the other clients, keeping the session consistent for everyone but
the client sending the erroneous request.