Add InfinotedTransformationProtector
This is a utility for infinoted that rejects requests that would require
transformation over many requests by other users. This allows to protect the
infinoted from users that are many requests behind and keep sending own
requests which would put a heavy load on the server.

The implementation is a bit hacky at this point. A cleaner solution would
be to add a query-execute-request signal to InfAdoptedAlgorithm, or to reject
the request on a lower level before execute-request is actually called.