Move transformation cache into InfAdoptedRequestLog
This is mainly an optimization of the algorithm cleanup routine.

Until now, every request made would trigger a traverse through the whole
request cache in order to see whether a request needs to be cleaned up
or not. This is optimized by instead of maintaining one big cache,
maintain for each user which only contains requests made by that user.
The requests are primarily sorted by the vector component of that user.

The actual cleanup is then done only when requests are removed from the
request log itself, instead of for processed request. Also, the tree is
only traversed until the first request is found that does not need to be
removed -- since the requests are sorted by the user component all other
requests would not be removed either.