Implement support for Access Control Lists (ACL)
This feature allows apply sets of permissions for different users on
individual nodes. Nodes which don't have an explicit setting for a permission
inherit the setting from their parent node. The default user for the root node
must have all permissions defined for the default user.

The user authentication currently works only with client side certificates.
This requires to run infinoted with the --ca-list-file option, since otherwise
clients are not requested to send a certificate. This behaviour needs yet to
be documented.

Users are identified by the SHA256 fingerprint of their certificate. This
means that, currently, a new certificate for a user would lead to a different
user account. This is not optimal. There should be fixed accounts to which one
or more certificates can be mapped, and which should also be able to be
associated to a SASL username, so that after successful SASL authentication
one has access to the account, without requiring client side certificates.

Right now, only two permissions are implemented, can-subscribe-session and
can-set-acl. With the can-subscribe-session permission certain users can be
allowed or denied to subscribe to particular sessions in the tree, and with
can-set-acl they can be allowed or denied to change the ACL for nodes. Of
course, more permissions will be added in the future. This commit is mostly
meant to provide the basic ACL framework, including ACL storage on the server
side, the API and the protocol layer to query, change and check ACLs.