Do not return a group that is larger than what can be undone in UndoGrouping
This could lead to submission of too many undo or redo requests if one would
blindly generate as many undo or redo requests as the size of the
corresponding undo group. Since it is not trivial to figure out the number of
requests allowed before no more requests can be made due to an excess of the
max total log size, this functionality has been added to
InfAdoptedUndoGrouping.

This would also lead to a crash in the server which cannot handle invalid
undo requests gracefully at the moment.