Watch notify::subscription-group instead of close for session detach
The reason is that when the session is closed, the subscription group is
reset as well in any case. This is being done by the default handler of
the InfSession::close signal. However, there are other occurences where
the subscription can be lost, for example if the server forcibly removes
the subscription with infd_session_proxy_unsubscribe(). If that happens,
we also want to detach the session from the browser, so that the client can
try to re-subscribe. The old session with lost subscription is kep alive,
however, so that it can be used for other purposes, such as for a sync-in
on another node.

The documentation of the InfBrowser::unsubscribe-session has also been
changed to describe its meaning more explicitly.