Operations in InfAdoptedSplitOperation are now one after each other
This defines the two operations in an InfAdoptedSplitOperation to be executed
one after each other, not anymore both operations at the same state. This
makes applying the split operation to the buffer straight forward, and no
internal concurrency ID would be needed to decide what to do when there are
two insert operations at the same position.

This fixes #610. Also, add a test and a replay reproducing the problem.