Use "forward translation" in InfAdoptedAlgorithm
This makes it possible to apply the mirror operation as late as possible
during a translation. Also, add a test case which failed before due to
this not being the case (test-55.xml).

This also simplifies the implementation of the adOPTed algorithm quite a bit.

At the moment the old implementation is still available and can be switched
back on. This is mostly for reference -- one of the new tests will fail with
the old implementation.

Unfortunately, there is also a performance impact with the new algorithm.
inf-test-text-session runs 30-40% slower. I hope this can be optimized more
later, and is another reason to keep the old algorithm still in place, so
that the two can easily be benchmarked against each other.

Also, replay-09.xml has been removed. This replay will fail with the new
algorithm. However, this is not an issue of the new algorithm, but rather
that the record was taken with the old algorithm, and therefore
transformations have been carried out incorrectly. The new algorithm carries
them out correctly, ends up in a different document state, and cannot play
the rest of the record.