Refactor infinoted option parsing
This was made having in mind introduction of generic infinoted plugins. The
parsing is now arranged such that in the end there is one GKeyFile from which
all options are read. Also, the configuration file to load can now be
specified on the command line, and the --help output is more useful.