Add generic plugin system
This allows to implement some special not-so-common functions as plugins,
mostly helping in keeping the core infinoted code clean and maintainable.

Features I have in mind for changing into a plugin are: autosave,
directory-sync, transformation-protector, session recording,
directory logging, traffic logging, and maybe also password, PAM
and/or certificate authentication.