Allow request functions to finish synchronously
This adds two additional arguments to all functions, which is the callback
function for the finished signal. The functions now work such that they
either return a request object with the given function connected to the
finished handler, or they return NULL in which case they must have called
the given function during their invocation.

This simplifies the implementation on the server side a lot, since requests
do not need to be artificially delayed any longer.

Also, using the API is a a bit simpler since in many cases one does not need
to use the request object anymore at all.