Remove the InfAdoptedAlgorithm::apply-request signal
This is a first step for proper error reporting from the algorithm. The
resulting operation might not be applicable to the buffer if the input was
invalid, and in order to report this the function should eventually be able
to report the error. However, this is problematic if it only runs as a
signal handler.

Therefore, the signal has been removed. The signal was used in three places.
In InfTextSession it was used to detect whether a buffer change was caused by
an executed request or externally. This use has been replaced by the new
inf_adopted_algorithm_get_execute_request() function.
inf_adopted_algorithm_execute_request() is not re-entrant anyway, so that the
introduction of this function does not cause a loss of generality.

The other two cases are in the libinfinity test suite where they can easily
be replaced by connecting to text-inserted and text-erased of the underlying
buffer instead.