Replace execute-request by begin-execute-request and end-execute-request
This allows to run the request execution not in the default signal handler
of the execute-request signal, which we need for error reporting from request
exeuction. However, we still need a possibility to hook into request
execution both before and after the request was transformed. Therefore the two
new signals have been introduced. This also allows to provide a different
set of parameters for the two; in end-execute-request we now provide the
result of the request execution, and also the reversible original request.

There is also an error parameter already which will be used later when error
reporting is implemented.