Fix a crash when receiving cursor updates in quick succession
In this case InfTextGtkView would install an idle handler for actual
determination of the new cursor and selection areas without checking
whether there was already an idle handler installed, leading to a
possible crash on the second occasion when the user has been set
inactive meanwhile, for example by the server unsubscribing the client.

This can be artifically provoked by Ctrl+Zing infinoted with enabled
transformation-protection plugin, making a couple of edits or cursor
movements in a 2nd gobby, then making another edit in the first gobby,
and then reviving the infinoted, so that it sends all updates immediately
to the first gobby and unsubscribing it directly afterwirds because its
own request exceeds the vdiff threshold of the transformation-protection
plugin.