Fix a crash when accepted connection is immediately closed by remote
In this case, getpeername() would fail, with an error of ENOTCONN. However,
the code assumed getpeername() would never fail.

This has been fixed in two ways now. First, the result of getpeername() is
checked for errors. Second, in the case of accepting a connection, we use
the result from accept() to find the peer's address, and no longer call
getpeername(). This way, we avoid an error during getpeername() failing, and
should receive a proper socket closure with the normal mechanism of send() or
recv() returning 0.

This bug was trigged by nmap scanning the infinoted port. Thanks to
Rainer Rehak <rainer@laryllian.de> for spotting this!