Make sure that browser is set when emitting "browser-set" signal
This affects InfGtkBrowserModelFilter and InfGtkBrowserModelSort. We now
connect to the browser-set signal of the child model after its default handler
is run, so that when we emit our own browser-set signal the browser is
actually accessible via the GtkTreeModel API.

The InfGtkBrowserView implementation is now also changed such that it is no
longer required that before the default handler run the old browser is still
available in the model. Before, it assumed that after the default handler ran
the new browser would be available, which was not true for
InfGtkBrowserModelFilter and InfGtkBrowserModelSort.

The semantic now is that, when connecting after the default handler of the
set-browser signal, the browser is set in the model. When connecting before
the default handler, the old or the new browser might be set in the model.
In order to make sure that both the old and the new browser are available
at all times both browsers are now parameters of the signal.