Add request parameters to most InfBrowser signals
This is basically to provide more convenience in the API. One could always
subscribe to the "begin-request" signal, and then connect to the "finished"
signal of the newly started request.

The request parameter can for instance be used to check whether the action was due
to some local request. In InfcBrowser this would basically be all actions that
have a request set. In InfdDirectory, this would be all actions whose request
does have the "requestor" property set to NULL.

A future commit will introduce a convenience API call for this check.