Add the certificate-auth plugin
This plugin enables client authentification with client certificates. The
functionality has already been available in the infinoted server itself and
is now moved into a plugin.

The has additional features compared to the previous functionality, including
the accept-unauthenticated-clients option and the feature to create a
short-lived super account that can be used to set up the initial permissions.