Fix a crash when server does not support a permission that the client does
The client would then mask out the root node permission of the
"default account", which is not allowed: the default account needs to have
the full mask set on the root node. This patch fixes the problem by making
sure the full mask is set in handle_set_acl. Unsupported permissions are then
assumed to correspond to the default ones.