InfGtkBrowserView: Allow initial root expansion in multiple views
With the change that initial root expansion is performed by default in
commit ef1ccbe5492e0d086e618ce323f3910b837bcf60, the current assumption
that there is at most one view per request that expands the root node
after initial exploration is no longer valid. The code is now changed such
that the root node can be expanded in multiple views at the same time.