Fix plugin manager cleanup when unloading a plugin causes unsubscriptions
The connection to the "unsubscribe-session" signal was removed too early, so
that the cleanup of the plugin manager session hash table was not properly
performed when a plugin's on_deinitialize() callback causes sessions to be
unsubscribed. This can happen when the "note-text" plugin is deinitialized on
server shutdown, and lead to an assertion failure at time of plugin manager
destruction.