Impose fewer constraints on the server, protocol-wise
For example, the server does no longer need to be able to provide a full
list of accounts, or to provide notifications when accounts are added or
removed.

This is preparation work for a new interface that allows to provide the
user accounts from an external source such as a database, LDAP, pam, or
other means.