infinoted: Make sure startup errors also show in log file
This allows system administrators to only filter the log file for
errors, and they don't have to keep also an eye on stderr.

However, early startup errors are still only logged to stderr, in case
they happen before the log file could be opened, for example when the
configuration file is incorrectly formatted.