Don't trust every certificate where the issuer is not found
The code was written under the assumption that when asking GnuTLS to
verify a certificate, it always fills the verification result fully.
However, when the issuer is not found, it immediately returns and does
not report other reasons why the certificate might be invalid.

However, we need to do at least a certificate verification even if the
signer of the certificate is not found, because in the latter case, we
use trust-on-first-use, and if the user decides so, still accept the
certificate. This is now fixed by running the certificate validation a
second time when the signer is not found, where we explicitly provide
the signer sent by the server.