Fix possible deadlock in dbus plugin
Apparently g_main_loop_quit() cannot be used from one thread to quit the
main loop in another thread. Instead, an idle source is added now.

Unfortunately, there is some strange crash happening inside a glib thread when
the plugin is unloaded. That thread seems to do some more deinitialization,
but we don't have a possibility to wait for it. We might need to use some more
low-level dbus API to fix that properly.