Fix crash after unloading dbus plugin
Loading the dbus plugin also loads libgio-2.0.so, and when the plugin is
unloaded, both are unloaded again. However, unloading libgio-2.0.so does
not really work because it registers static memory with the gobject type
system. Therefore, make sure that libgio stays loaded even after the dbus
plugin gets unloaded.