Make sure to prefer DHE-RSA over RSA
Recent versions of GnuTLS prefer RSA over DHE-RSA for comatibility
reasons. If a server negotiates a DH group with a low number of prime
bits, the handshake will inevitably fial and there is no way to fall
back to RSA. However, the DH groups that Gobby and infinoted generate
have 2048 bits, so we don't expect this to be a problem in practise,
and can therefore benefit from DHE-RSA. This might become a problem
in the future, and we can re-visit then.

Note that when Gobby and infinoted are used with recent versions of
GnuTLS, they also prefer ECDHE over DHE, in which case this is not
an issue anyway.

This change basically fixes PFS when a client with a recent GnuTLS
version connects to a server with a not-so-recent GnuTLS version
(such as 2.12). We can do this because we know that the DH group that
that server will have is strong enough.