InfTextChunk: fix segment lookup for offset=0 (#10)
This used to work with glib 2.42, but it seems that the semantics of
g_sequence_search() have changed with respect to what item is returned
when the comparison function returns 0. The behavior in that case is not
documented. Fix this by passing a different comparison function that
never returns 0, so that there is no ambiguity in which segment is returned.