Add auto-detected support for libsystemd.
systemd has a service readiness protocol in which the daemon sends a message
over a Unix socket when it set up the listeners and is ready to serve.
libsystemd abstracts this away to be a simple call to sd_notify. Hence
hook up infinoted to signal systemd just before it enters the I/O main loop.

Signed-off-by: Philipp Kern <phil@philkern.de>